refactor(project): rename packages & add required keys
