+++ src/controllers/song-controller.js
@@ -1,4 +1,4 @@
-const Song = require('../model/song')
+const Song = require('../models/song')
 
 exports.findAll = function * () {
     this.body = yield Song.find({})
