+++ src/index.js
@@ -2,10 +2,10 @@ const koa = require('koa')
 const logger = require('koa-logger')
 const route = require('koa-route')
 
-const db = require('./util/db')
-const config = require('./util/config')
-const constants = require('./util/constants/path-constants')
-const songController = require('./controller/song-controller')
+const db = require('./utils/db')
+const config = require('./utils/config')
+const constants = require('./utils/constants/path-constants')
+const songController = require('./controllers/song-controller')
 
 const server = module.exports = koa()
 server.name = config.get('app:name')
